Moved file locations and cleaned up console logs
