--- server_node/packages/pumpfun/example/basic/.keys/test-account.json
@@ -1,1 +1,0 @@
-{"secretKey":"2UY4i4jtVFvyBGzcLcuuUN2jbvQYh5SrGgBTq1Z4yWu5o4EXJvZBLLJAi8V7nUUjXXEwX15QNzEoepDgn7G13bZ2","publicKey":"HFTKPREjYoVCbmGKJ6i2bjwXAvWR4Bsq1PTkyTduy5rG"}
